double -> int

+++ core-lib/es5/src/main/java/def/dom/TouchList.java
@@ -1,8 +1,8 @@
 package def.dom;
 public class TouchList extends def.js.Object implements Iterable<Touch> {
-    public double length;
-    native public Touch item(double index);
-    native public Touch $get(double index);
+    public int length;
+    native public Touch item(int index);
+    native public Touch $get(int index);
     public static TouchList prototype;
     public TouchList(){}
     /** From Iterable, to allow foreach loop (do not use directly). */
