+++ core-lib/es6/src/main/java/def/dom/TouchList.java
@@ -4,9 +4,9 @@ import def.js.Iterable;
 
 @jsweet.lang.SyntacticIterable
 public class TouchList extends Iterable<Touch> {
-    public double length;
-    native public Touch item(double index);
-    native public Touch $get(double index);
+    public int length;
+    native public Touch item(int index);
+    native public Touch $get(int index);
     public static TouchList prototype;
     public TouchList(){}
     /** From Iterable, to allow foreach loop (do not use directly). */
